create class for convert_ac_line_segments_cim16
